Move existing functional tests to test/functional/
